renamed `diafiltration_two_salt.py` to `multi_component_diafiltration.py`
