+++ src/prommis/nanofiltration/diafiltration_flowsheet_two_salt.py
@@ -30,7 +30,7 @@ from prommis.nanofiltration.diafiltration_stream_properties import (
     DiafiltrationStreamParameter,
 )
 from prommis.nanofiltration.diafiltration_solute_properties import SoluteParameter
-from prommis.nanofiltration.diafiltration_two_salt import TwoSaltDiafiltration
+from prommis.nanofiltration.multi_component_diafiltration import TwoSaltDiafiltration
 
 
 def main():
